backend/middleware_logging.py
import contextvars
import logging
import time
import uuid

from starlette.middleware.base import BaseHTTPMiddleware
from starlette.requests import Request
from starlette.responses import Response

logger = logging.getLogger(__name__)

request_id_var = contextvars.ContextVar("request_id", default="-")
user_uid_var = contextvars.ContextVar("user_uid", default="-")


class RequestContextMiddleware(BaseHTTPMiddleware):
    async def dispatch(self, request: Request, call_next) -> Response:
        req_id = uuid.uuid4().hex[:8]
        request_id_var.set(req_id)
        start = time.time()
        response = await call_next(request)
        duration_ms = round((time.time() - start) * 1000)
        # Skip noisy health check logs
        if request.url.path != "/health":
            logger.info("request_complete", extra={
                "request_id": req_id,
                "method": request.method,
                "path": request.url.path,
                "status": response.status_code,
                "duration_ms": duration_ms,
                "user_uid": user_uid_var.get(),
            })
        return response
