backend/requirements.txt
fastapi>=0.124.1,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
google-adk==1.25.0
google-genai==1.64.0
google-cloud-firestore>=2.19.0,<3.0.0
google-cloud-storage>=2.18.0,<3.0.0
httpx>=0.27.0,<1.0.0
beautifulsoup4==4.12.3
pydantic>=2.12.0,<3.0.0
python-dotenv==1.0.0
python-multipart>=0.0.9,<1.0.0
Pillow>=10.0.0,<11.0.0
resend==2.0.0
sse-starlette>=1.8.2,<3.0.0
cryptography>=42.0.0,<44.0.0
firebase-admin>=6.5.0,<7.0.0
python-json-logger==3.2.1
# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
